fix: 責務の分離
