+++ virtual_partner_front/src/features/Niconico/const.ts
@@ -1,0 +1,30 @@
+export const mockComments = [
+  "実家帰るんか？マッマ泣いてるぞ",
+  "金ないなら近所の公園でええやろ",
+  "ええなぁ、ワイはバイト漬けや…",
+  "ぼっちがどこ行くんやw",
+  "海行こうぜ、ナンパできんけど",
+  "コスパ最強、図書館で過ごせ",
+  "真夏の甲子園やろ！熱気ヤバいで",
+  ">>1、彼女おるん？おらんなら家でゲームや",
+  "お前ら揃ってBBQしようぜwww（絶対来ないやつ）",
+  "夏休みは引きこもりが正解や",
+  "地方の温泉ええぞ、安いし快適や",
+  "キャンプで虫刺されて泣く未来しか見えん",
+  "ワイのオススメ、昼間のショッピングモールや",
+  "花火大会行く奴はリア充認定や",
+  "避暑地とか言いながらカフェ巡りするんやろ",
+  "ワイ、マジでどこも行く予定ないわ…",
+  "旅行先の飯だけはちゃんと調べとけよ",
+  "お前らまた帰りの新幹線で泣くんやろ",
+  "夏休みは昼夜逆転から始まるで",
+  "アニメの聖地巡礼とか興味ないん？",
+  "バーベキューとかしたいけど準備がな…",
+  ">>1、お土産は金沢のゴリラパンな",
+  "夏休みって言ってもお前には毎日が夏休みやん",
+  "オタク特有の秋葉原行っとけばいいって考えやめろ",
+  "家で冷房最強にしてアイス食べるのが勝ち組",
+  "コミケ参戦の準備はええんか？",
+  "電車賃すらないワイ、高みの見物",
+  "ワイ、どうせ今年もお盆休みゼロやで"
+];
